Finished inital implementation of parallel. Tests not passing.
